scripts to test webpack public path
The build:top-test script builds the project with the DEPLOY_PATH variable
and sets the output folder to top-test/specific/release
and then copies the index-top.html file to the top-test/

The serve:top-test then runs a webserver in top-test/
So you can make sure the app works when built this way.

I'm pretty sure we could setup a way to test this with the webpack dev server
but this build approach is exactly how it will be handled in production so it is
a better integration test.